New config: Start flashMode and remove FlasjMode button - none breaking.
This adds the option to configure the camera to start in another flash mode than auto.

It also adds the possibility to remove the flash cycle button from the UI. Typically used when FlashMode.off is used and it is not desired to use any other mode.

The cycle and mode changing logic was updated to a version that could pass the linter and also enabled no showing the flash mode while camera changes from default mode, to desired mode.

A local state variable _flashMode was introduced to manage the flash mode state.